+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [next]
 
 - Japanese Kanji support added using [`kuromoji`](https://pub.dev/packages/kuromoji).
+- Hebrew romanization support added based on ISO 259 / DIN 31636.
 
 ## 0.0.1